refactor: Auth UI 파일 통합
