--- frontend/src/features/auth/ui/AuthBottomContainer.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-const AuthBottomContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 40px;
-  text-align: center;
-`;
-
-export default AuthBottomContainer;
