--- frontend/src/features/auth/ui/AuthButtonContainer.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const AuthButtonContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 20px;
-  width: 100%;
-`;
-
-export default AuthButtonContainer;
