--- frontend/src/features/auth/ui/AuthForm.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const AuthForm = styled.form`
-  width: 100%;
-`;
-
-export default AuthForm;
