--- frontend/src/features/auth/ui/AuthWrapper.ts
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-const AuthWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100svh;
-
-  & > div {
-    width: 100%;
-  }
-`;
-
-export default AuthWrapper;
